+++ stories/components/Itinerary.stories.js
@@ -19,7 +19,8 @@ import sampleItinerary from '@/assets/fixtures/itineraries/sample-itinerary.json
 
 export default {
     component: VsItinerary,
-    title: 'Components/To be reviewed/Itinerary',
+    title: 'Components/Deprecated/Itinerary',
+    tags: ['deprecated'],
 };
 
 const Template = (args) => ({
@@ -47,6 +48,11 @@ const Template = (args) => ({
         };
     },
     template: `
+        <div style="background: #940f2c17; padding: 1rem; border: 1px solid #940F2C; margin-bottom: 1rem;">
+            ⚠️ Deprecated: Use new article page layouts and guidance. This component will be removed in a future release.
+            See the <a href="https://visitscotland.atlassian.net/wiki/spaces/DE/pages/1273397342/Itinerary+deprecated">Component migration docs.</a>
+        </div>
+
         <div
             :class="args.jsDisabled ? 'no-js' : ''"
         >